data: split homepage-layout.js
