+++ src/data/things/homepage-layout/index.js
@@ -1,0 +1,6 @@
+export * from './HomepageLayout.js';
+export * from './HomepageLayoutSection.js';
+export * from './HomepageLayoutRow.js';
+export * from './HomepageLayoutActionsRow.js';
+export * from './HomepageLayoutAlbumCarouselRow.js';
+export * from './HomepageLayoutAlbumGridRow.js';
