+++ src/data/things/index.js
@@ -1,5 +1,7 @@
 // Not actually the entry point for #things - that's init.js in this folder.
 
+export * from './homepage-layout/index.js';
+
 export * from './AdditionalFile.js';
 export * from './AdditionalName.js';
 export * from './ArtTag.js';
@@ -17,5 +19,4 @@ export * from './album.js';
 export * from './content.js';
 export * from './flash.js';
 export * from './group.js';
-export * from './homepage-layout.js';
 export * from './sorting-rule.js';
